+++ fixtures/app_with_plugin/Source/AppDelegate.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let viewController = UIViewController()
+        viewController.view.backgroundColor = .white
+        viewController.view.addSubview(TestView())
+        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
+        return true
+    }
+}