+++ fixtures/app_with_plugin/Tuist/Config.swift
@@ -1,0 +1,9 @@
+import ProjectDescription
+
+let config = Config(
+    plugins: [
+        .local(path: .relativeToManifest("../../LocalPlugin")),
+        .git(url: "https://github.com/tuist/ExampleTuistPlugin.git", tag: "1.0.0")
+    ],
+    generationOptions: []
+)